Attempt to a more generalized support for unions
  G |- S <: T1 + ... + Tn

would get broken down into the following:

  G,     K1                              |- S <: T1
  G, not K1, K2                          |- S <: T2
  ...
  G, not K1, ..., not K(n-2),     K(n-1) |- S <: T(n-1)
  G, not K1, ..., not K(n-2), not K(n-1) |- S <: Tn

where K1, ..., K(n-1) need only be instantiated with true or false.
At least one line of the n will not be trivially true.

This doesn't work ATM because the K-vars at the LHS of a constraint
are only assigned once -- there is no backtracking to check other
values if the first assignment fails.